// types/database.ts
export type Database = {
  public: {
// ...
    Enums: {
      booking_status:
        | "pending"
        | "confirmed"
        | "cancelled"
        | "completed"
        | "no_show"
      business_status: "pending" | "approved" | "rejected" | "inactive"
      business_type: "accommodation" | "shop" | "service"
      content_status: "pending" | "approved" | "rejected"
      content_type: "business_profile" | "tourist_spot" | "event" | "promotion"
      event_status: "upcoming" | "ongoing" | "completed" | "cancelled"
      page_view_type: "business" | "tourist_spot" | "event" | "promotion"
      payment_method: "gcash" | "paypal" | "xendit" | "credit_card" | "cash"
      payment_status:
        | "pending"
        | "paid"
        | "failed"
        | "refunded"
        | "partially_refunded"
      promotion_status: "active" | "scheduled" | "expired" | "cancelled"
      review_type: "business" | "tourist_spot" | "event"
      tourist_spot_status:
        | "active"
        | "inactive"
        | "under_maintenance"
        | "coming_soon"
      tourist_spot_type:
        | "natural"
        | "cultural"
        | "historical"
        | "religious"
        | "recreational"
        | "other"
      user_role:
        | "tourism_admin"
        | "business_listing_manager"
        | "tourism_content_manager"
        | "business_registration_manager"
        | "business_owner"
        | "tourist"
    }
    CompositeTypes: {
      geometry_dump: {
        path: number[] | null
        geom: unknown | null
      }
      valid_detail: {
        valid: boolean | null
        reason: string | null
        location: unknown | null
      }
    }
  }
}

type DefaultSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never